+++ mobile/android/android-components/samples/browser/src/main/java/org/mozilla/samples/browser/BrowserFragment.kt
@@ -128,8 +128,9 @@ class BrowserFragment : Fragment(), BackHandler {
 
         sitePermissionsFeature.set(
             feature = SitePermissionsFeature(
-                anchorView = layout.toolbar,
-                sessionManager = components.sessionManager
+                context = requireContext(),
+                sessionManager = components.sessionManager,
+                fragmentManager = requireFragmentManager()
             ) { permissions ->
                 requestPermissions(permissions, REQUEST_CODE_APP_PERMISSIONS)
             },